Added a ChargingStation object variable and getChargingStation() method to the robot class.
new Orchestrator method getRobotsOrderedByDistance finds all the available robots, sorts them by distance ascending to the specified position and returns this as an arraylist of map entries (Robot->Distance)

Warehouse now has floorPlanX and floorPlanY variables and getMethods for these.

Updated the route in Job class to be an LinkedList of linked lists so that the total route can be broken down into segments, i.e go to shelf, go to packing station, go to charging station.
Changed Job class to include an LinkedList of actions. The idea is to have "Collect" "Drop" and "Charge" as the actions, and have the robot read these and remove them from the LinkedList as it completes parts of the route.

Added some functionality to the Entity constructor to reject an invalid position and stop the simulation.