Orchestrator now stores the warehouse object.
Tested route creation for a single robot round trip. Added some prints in the createJobs method to print the routes.

Added a new constructor for the Job class that assigns default actions.

Added a getRoutes method to the Job class, which returns all routes stored for that job.

Updatd the getNextRoute method, which now utilises the iterator of the LinkedList.

Defined a custom exception for ChargingStation that gets thrown when trying to assign a robot to a charging station that already has a robot.

Defined custom exceptions for the order class to handle invalid IDs and number of items. Associated logic added where required in the constructor.

Defined a new method for the Charging Station, setRobot, which takes in a robot object and assigns it to the robot. This is so the ChargingStation can increase the charge of its robot and allow other robots to inhabit the same tile when it is free, in case a route passes over the charging station.

Fixed the very broken LinkedList custom class. Implemented Iterable and created a LinkedListIterator class so it can be iterated over nicely using a foreach loop. Think this is in good condition but further testing is required.

Jobs now get assigned to robots but robots do nothing yet.

Added a toString method to the Position class which prints out the coordinates of the position.